+++ deflect/Event.h
@@ -86,7 +86,8 @@ struct Event
         EVT_KEY_RELEASE,
         EVT_VIEW_SIZE_CHANGED,
         EVT_TAP_AND_HOLD,
-        EVT_PAN
+        EVT_PAN,
+        EVT_PINCH
     };
 
     /** The type of event */
@@ -96,8 +97,8 @@ struct Event
     //@{
     double mouseX;    /**< Normalized X mouse/touch position relative to the window */
     double mouseY;    /**< Normalized Y mouse/touch position relative to the window */
-    double dx;        /**< Normalized horizontal delta for scroll events / delta in pixels for wheel events. */
-    double dy;        /**< Normalized vertical delta for scroll events / delta in pixels for wheel events. */
+    double dx;        /**< Normalized horizontal delta for pan/pinch events / delta in pixels for wheel events. */
+    double dy;        /**< Normalized vertical delta for pan/pinch events / delta in pixels for wheel events. */
     bool mouseLeft;   /**< The state of the left mouse button (pressed=true) */
     bool mouseRight;  /**< The state of the right mouse button (pressed=true) */
     bool mouseMiddle; /**< The state of the middle mouse button (pressed=true) */